% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{Shannon diversity}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector of numeric values representing the species abundance in a community}
}
\value{
Shannon diversity value
}
\description{
This is a shannon function
}
\examples{
com <- c(1, 4, 7, 8)
my_shannon(com)
}
